updated top populated cities in the region (fixed data type)

+++ src/main/java/com/napier/coursework/City.java
@@ -38,7 +38,7 @@ public class City {
         return District;
     }
 
-    public double getPopulation() {
+    public int getPopulation() {
         return Population;
     }
 }
